requirements.txt
replicate>=0.26,<1
click>=8.1,<9
python-dotenv>=1.0,<2
requests>=2.31,<3
tqdm>=4.66,<5
tenacity>=8.2,<9

